bot/requirements.txt
aiogram==3.0.0b3
environs~=8.0.0
python-dateutil
loguru